Move project admin form to forms.py
